refactor(background): move scheduled job to correct module
